Enable option updates
Updating options via script-opts-append already worked,
however it required to actually have the option in script-opts in order
to change it's value. That means it was e.g. not possiblet to remove an
option from script-opts to fall back to the value configured in the
.conf file.

By enabling option updates the script-opts always get applied on top of
the original configuration, which means removing an option from
script-opts now also restores the original value.

This removes the need for setting default script-opts when changing
options via conditional profiles.